Fixed package.json description

+++ src/NewCommand.php
@@ -1,16 +1,35 @@
 <?php
+/**
+ * Generate a new installation of AWPS
+ *
+ * @category CLI
+ * @package  Awps-cli
+ * @author   Alessandro Castellani <me@alecaddd.com>
+ * @license  http://opensource.org/licenses/gpl-3.0 GNU General Public License, version 3 (GPLv3)
+ * @link     http://alecaddd.com
+ */
 
 namespace Awps;
 
+use ZipArchive;
+use RuntimeException;
+use GuzzleHttp\ClientInterface;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use GuzzleHttp\ClientInterface;
 use Symfony\Component\Console\Question\Question;
-use ZipArchive;
 
+/**
+ * Generate a new installation of AWPS
+ *
+ * @category CLI
+ * @package  Awps-cli
+ * @author   Alessandro Castellani <me@alecaddd.com>
+ * @license  http://opensource.org/licenses/gpl-3.0 GNU General Public License, version 3 (GPLv3)
+ * @link     http://alecaddd.com
+ */
 class NewCommand extends Command
 {
 	private $client;
@@ -22,34 +41,62 @@ class NewCommand extends Command
 		parent::__construct();
 	}
 
+	/**
+	 * Configure the CLI command
+	 *
+	 * @return void
+	 */
 	public function configure()
 	{
-		$this->setName('new')
-		->setDescription('Create a new AWPS Theme installation')
-		->addArgument('name', InputArgument::REQUIRED, 'Insert the folder name');
+		$this
+			->setName('new')
+			->setDescription('Create a new AWPS Theme installation')
+			->addArgument('name', InputArgument::REQUIRED, 'Insert the folder name')
+			->addOption('dev', null, InputOption::VALUE_NONE, 'Download the latest "development" release')
+			->addOption('force', null, InputOption::VALUE_NONE, 'Forces the install even if the directory already exists');
 	}
 
+	/**
+	 * Execute the creation command
+	 *
+	 * @param InputInterface $input
+	 * @param OutputInterface $output
+	 * @return void
+	 */
 	public function execute(InputInterface $input, OutputInterface $output)
 	{
-		$themeName = $input->getArgument('name');
-		$directory = getcwd().'/'.$themeName;
+		if (! class_exists('ZipArchive')) {
+			throw new RuntimeException('The Zip PHP extension is not installed. Please install it and try again.');
+		}
+
+		$folderName = $input->getArgument('name');
+		$directory = getcwd().'/'.$folderName;
 
 		$output->writeln('<info>Summoning application..</info>');
 
-		$this->assertApplicationDoesNotExists($directory, $output);
+		if (! $input->getOption('force')) {
+			$this->assertApplicationDoesNotExists($directory, $output);
+		}
 
 		$this->assertLocationInsideWordPress($directory, $output);
 
 		$helper = $this->getHelper("question");
-		$question = new Question("Change Namespace? <info>(Awps)</info> ", "Awps");
+
+		$question = new Question("Name of your theme? <info>(Awps)</info> ", null);
+		$themeName = $helper->ask($input, $output, $question);
+
+		$question = new Question("PHP Namespace of your theme? <info>(Awps)</info> ", null);
 		$namespace = $helper->ask($input, $output, $question);
+
+		$question = new Question("Description? <info>(Alecaddd WordPress Starter theme)</info> ", null);
+		$description = $helper->ask($input, $output, $question);
 
 		$output->writeln('<info>Downloading Package..</info>');
 
 		$this->download($ZipFile = $this->makeFileName())
-		->extract($ZipFile, $directory, $output)
-		->renameNamespaces($directory, $output, $themeName, $namespace)
-		->cleanUp($ZipFile);
+				->extract($ZipFile, $directory, $output)
+				->renameAllTheThings($directory, $output, $themeName, $namespace, $description)
+				->cleanUp($ZipFile);
 
 		$output->writeln('<info>Updating config files..</info>');
 
@@ -59,6 +106,13 @@ class NewCommand extends Command
 		$output->writeln('<comment>Application ready, Happy Coding!</comment>');
 	}
 
+	/**
+	 * Assert if the folder doesn't already exists
+	 *
+	 * @param string $directory
+	 * @param OutputInterface $output
+	 * @return void
+	 */
 	private function assertApplicationDoesNotExists($directory, OutputInterface $output)
 	{
 		if (is_dir($directory)) {
@@ -67,19 +121,37 @@ class NewCommand extends Command
 		}
 	}
 
+	/**
+	 * Assert the cli was triggerd inside a WordPress directory
+	 *
+	 * @param string $directory
+	 * @param OutputInterface $output
+	 * @return void
+	 */
 	private function assertLocationInsideWordPress($directory, OutputInterface $output)
 	{
-		if (!strpos(dirname($directory), 'wp-content')) {
+		if (! strpos(dirname($directory), 'wp-content')) {
 			$output->writeln('<error>This is not a WordPress theme directory!</error>');
 			exit(1);
 		}
 	}
 
+	/**
+	 * Generate a temporary file name for the downloaded zip file
+	 *
+	 * @return void
+	 */
 	private function makeFileName()
 	{
 		return getcwd().'/awps_'.md5(time().uniqid()).'.zip';
 	}
 
+	/**
+	 * Download the lates AWPS release
+	 *
+	 * @param string $zipFile
+	 * @return void
+	 */
 	private function download($zipFile)
 	{
 		$response = $this->client->get('https://github.com/Alecaddd/awps/archive/master.zip')->getBody();
@@ -89,6 +161,14 @@ class NewCommand extends Command
 		return $this;
 	}
 
+	/**
+	 * Extract the downlaod zip
+	 *
+	 * @param string $ZipFile
+	 * @param string $directory
+	 * @param OutputInterface $output
+	 * @return void
+	 */
 	private function extract($ZipFile, $directory, OutputInterface $output)
 	{
 		$archive = new ZipArchive();
@@ -99,7 +179,7 @@ class NewCommand extends Command
 
 		$archive->close();
 
-		if (!empty(shell_exec("mv ".$directory."/awps-master/* ".$directory."/awps-master/.[!.]*  ".$directory))){
+		if (! empty(shell_exec("mv ".$directory."/awps-master/* ".$directory."/awps-master/.[!.]*  ".$directory))) {
 			$output->writeln('<comment>Unable to move files from master folder!</comment>');
 
 			return $this;
@@ -110,24 +190,46 @@ class NewCommand extends Command
 		return $this;
 	}
 
-	private function renameNamespaces($directory, OutputInterface $output, $themeName = null, $namespace = null)
+	/**
+	 * Rename the application PHP namespace
+	 *
+	 * @param string $directory
+	 * @param OutputInterface $output
+	 * @param string $themeName
+	 * @param string $namespace
+	 * @param string $description
+	 * @return void
+	 */
+	private function renameAllTheThings($directory, OutputInterface $output, $themeName = null, $namespace = null, $description = null)
 	{
-		if(is_null($themeName) && is_null($namespace)) return $this;
+		if (is_null($themeName) && is_null($namespace) && is_null($description)) {
+			return $this;
+		}
+
+		$output->writeln('<info>Updating namespaces..</info>');
 
 		$file_info = array();
 
 		$this->recursiveScanFiles($directory, $file_info);
 
-		foreach($file_info as $file) {
+		foreach ($file_info as $file) {
 			$str = file_get_contents($file);
 
-			if(!is_null($namespace)) {
-				$str = str_replace( "Awps", $namespace, $str );
-				$str = str_replace( "awps", $namespace, $str );
+			if (! is_null($namespace)) {
+				$str = str_replace("awps", $namespace, $str);
+				$str = str_replace("use Awps", "use " . $namespace, $str);
+				$str = str_replace("namespace Awps", "namespace " . $namespace, $str);
+				$str = str_replace("Awps\\", $namespace . "\\", $str);
+				$str = str_replace("Awps\Init", $namespace . "\Init", $str);
+			}
+
+			if (! is_null($themeName)) {
+				$str = str_replace("Awps", $themeName, $str);
 			}
 			
-			if(!is_null($themeName)) {
-				$str = str_replace( "Alecaddd WordPress Starter theme", $themeName, $str );
+			if (! is_null($description)) {
+				$str = str_replace("Alecaddd WordPress Starter theme", $description, $str);
+				$str = str_replace("Alecaddd WordPress Starter Theme for savvy developers", $description, $str);
 			}
 
 			file_put_contents($file, $str);
@@ -136,13 +238,15 @@ class NewCommand extends Command
 		return $this;
 	}
 
-	private function recursiveScanFiles($path, &$file_info){
+	private function recursiveScanFiles($path, &$file_info)
+	{
 		$path = rtrim($path, '/');
-		if(!is_dir($path)) $file_info[] = $path;
-		else {
+		if (! is_dir($path)) {
+			$file_info[] = $path;
+		} else {
 			$files = scandir($path);
-			foreach($files as $file) {
-				if($file != '.' && $file != '..') {
+			foreach ($files as $file) {
+				if ($file != '.' && $file != '..') {
 					$this->recursiveScanFiles($path . '/' . $file, $file_info);
 				}
 			}
@@ -159,15 +263,19 @@ class NewCommand extends Command
 
 	private function moveConfig($directory, OutputInterface $output)
 	{
-		if (!empty(shell_exec("mv ".$directory."/wp-config.sample.php ../../"))) {
+		if (! empty(shell_exec("mv ".$directory."/wp-config.sample.php ../../"))) {
 			$output->writeln('<comment>Unable to move the wp-config.sample.php file, be sure to move it in the base directory!</comment>');
 		}
 
-		if (!empty(shell_exec("mv ".$directory."/.env.example ../../.env"))) {
+		if (is_file("../../.env")) {
+			$output->writeln('<comment>Existing .env detected in the WordPress root directory, nothing to do here!</comment>');
+			return $this;
+		}
+
+		if (! empty(shell_exec("mv ".$directory."/.env.example ../../.env"))) {
 			$output->writeln('<comment>Unable to move the .env.example file, be sure to move it in the base directory and rename it as .env!</comment>');
 		}
 
 		return $this;
 	}
-
 }
